Change inputs to float for decimal calculations

+++ Calculator.py
@@ -1,6 +1,6 @@
 print("   Simple Calculator    ")
-num1=int(input("Enter the first number= "))
-num2=int(input("Enter the second number= "))
+num1=float(input("Enter the first number= "))
+num2=float(input("Enter the second number= "))
 
 opr=input("Select operator(+ - * /)= ")
 
@@ -18,5 +18,6 @@ elif opr==("/"):
 else:
     print("Invalid operator or number")
     
+
